+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Allow custom blocks in `String#tokenizeHTML()` to drop their delimiters from the output
 * Add line_start & line_end to String.tokenizeHTML() output
+* Add line_start & line_end to Function.tokenize() output
 
 ## 0.7.12 (2021-08-10)